pc/pcmciamodem: cast i/o port to ulong

+++ sys/src/9/pc/pcmciamodem.c
@@ -68,9 +68,9 @@ pcmciamodemlink(void)
 				com2used = 1;
 			if(ioalloc(isa.port, 8, 0, modems[j]) < 0)
 				print("%s port %lux already in use\n",
-					modems[j], isa.port);
+					modems[j], (ulong)isa.port);
 			print("%s in pcmcia slot %d port 0x%lux irq %d\n",
-				modems[j], slot, isa.port, isa.irq);
+				modems[j], slot, (ulong)isa.port, isa.irq);
 		}
 	}
 }
